# api/swagger.yaml
openapi: 3.0.0
info:
  title: Personal Finance Manager API
  description: API documentation for the Personal Finance Manager application
  version: 1.0.0
  contact:
    name: Backend Team
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [income, expense, transfer]
        icon:
          type: string
        color:
          type: string
        parent_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Budget:
      type: object
      properties:
        id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        spent:
          type: number
          format: double
        period:
          type: string
          enum: [monthly, quarterly, yearly]
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    NotificationPreferences:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        email_notifications:
          type: boolean
        push_notifications:
          type: boolean
        notification_types:
          type: array
          items:
            type: string
            enum: [low_balance, bill_due, budget_exceeded, unusual_activity]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
paths:
...
  /api/categories:
    get:
      summary: Get all categories
      tags: [Categories]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      summary: Create a new category
      tags: [Categories]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum: [income, expense, transfer]
                icon:
                  type: string
                color:
                  type: string
                parent_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
...
  /api/budgets:
    get:
      summary: Get user budgets
      tags: [Budgets]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Budget'
    post:
      summary: Create a new budget
      tags: [Budgets]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - category_id
                - amount
                - period
              properties:
                category_id:
                  type: string
                  format: uuid
                amount:
                  type: number
                  format: double
                period:
                  type: string
                  enum: [monthly, quarterly, yearly]
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
      responses:
        '201':
          description: Budget created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
...
  /api/notifications/preferences:
    get:
      summary: Get notification preferences
      tags: [Notifications]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User's notification preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferences'
    put:
      summary: Update notification preferences
      tags: [Notifications]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email_notifications:
                  type: boolean
                push_notifications:
                  type: boolean
                notification_types:
                  type: array
                  items:
                    type: string
                    enum: [low_balance, bill_due, budget_exceeded, unusual_activity]
      responses:
        '200':
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferences'